requirements.txt
gunicorn
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
line-bot-sdk
markupsafe==2.0.1
investpy